Rename Effect to QAbstractEffect
It appears that to easily create custom aspects without having to rely on the
RendererAspect, the core API needs to be extended to contain interfaces of
abstracts for Effect, Mesh, Material, Technique. This is the first patch of a
serie of patches to follow that will rename classes correctly prior to moving
them to Qt3D Core.

Change-Id: I6a6a5c8a7476875cd7be04248aebe747c217396a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>