Properly fix camera controller
The problem was that the vector from the camera world position to the
view center world position was not being updated when the camera moved.
This resulted in incorrect calculation of the worldspace movement vector
on subsequent frames. This meant adding a new setter on the LookAt
transform.

Also updated the tilt angle to follow the mouse rather than being
inverted like a flight sim.

Furthermore, the view matrix was incorrectly being inverted by the
RenderView.

These mistakes all combined led to camera behavior which was close
but not quite correct originally. This should now all be mathematically
correct and function as expected.

Change-Id: I85d27b37c7e6d5504d2ed4073b8ab2e4c7f04821
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>