QAbstractShader class
It might be interesting to abstract the shaders as well. This offers a way
to have a complete scene described and use the shaders for other things than
3d rendering, sound rendering for instance.

Each aspect can then check for each QAbstractShader in the scene if it is a
shader relevant to its context by performing a qobject_cast.

Change-Id: I645bba0536e88681ad21cdb2d7dd38661059e7bb
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>