Monitor Nodes for Creation/Deletion
Node is now a QObservable which is needed to tell the QChangeArbiter
when child nodes are appended or removed.
RenderNode is registered as an observer off the frontend Node it receives.

Note: QtQuick Repeater seems to rely on the delegate to be a QuickItem
subclass. It might not work with Qt3D Node. Using a Loader loading a Node
subtree from a QML file, the file is loaded but the subtree is not appended to
the root Scenegraph tree. I believe this is due to the fact that the loaded
file is set as the child of the Loader where we would want it to be set as the
child of the Node holding the Loader.

Change-Id: I9fa4e29bdbd7fdf51d5e0bd8e0a33c369ab78d90
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>