Use enum for OpenGL Types in Parameter
This offers a way not to have to include any related OpenGL header at that
point. That way we can, depending on the enum's value and the types available
when rendering determine the correct GL types.

Note: It seems the GLTF as evolved or that the parser isn't correct as it expects
raw GL values instead of String for types which contradicts the current
specifications.

Change-Id: I78a1a70d1dac429ab6bad23aacdac48a16cd1025
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>