Avoid race condition when updating view matrix
Also add check on property name for the view matrix update.

This change fixes the occasional visual glitch when moving the camera
rapidly. The glitch was caused by the main thread updating the view
matrix as the job worker threads were reading it. To avoid locking we
use the standard QChangeArbiter route.

Change-Id: I2cd4a44de8f1df0e6b944710239656f697bb3235
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>