All Node subclasses now use same route to talk to QChangeArbiter
Node no longer inherits from QObservable as QObservable is meant for
use only by aspects.

Node stores a QChangeArbiter pointer directly so that we can call the
dedicated locking functions on it. The locking is needed for Node
subclasses as these live in the main thread which is not managed by
the QAspectmanager to work in sync with the QChangeArbiter.

Change-Id: I1151ca107b4cb29dd303371c0095c022c43bc008
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>