ParameterBinder element
This element offers a way to specify the binding of parameters with a Shader
program in a RenderPass. A parameter is referenced by its name, the
shaderVariable by its name and the binding type has to be set to determine
if the binding is for an Attribute or a Uniform.

Change-Id: I75cdc9f760013f8f08aed82429afb6800305627c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>