Rendering now uses Material, Effect, Technique and Shaders
The Renderer has been updated to properly set uniforms and attributes based on
the values stored in each RenderCommand. RenderView constructs a QUniformPack
for non standard uniforms, StandardUniforms are stored in the RenderShader and
Attributes bindings are stored in the RenderCommand.

Follow-up patches will further improve the current code once we can remove and
clean areas that were set up for the hacked renderer.

Note: This breaks rendering of meshes relying on the default material. This
will be fixed in a separate patch.

Change-Id: I78f2a086cdc38ee9b403eeede5d290de9c07d2a8
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>