Refactoring of Renderer, RenderMaterial and RenderTechnique
Removed most of the methods that were needed by the hacked Renderer.
The default technique has also been modified to use ParameterBinder.
The Renderer now has a simpler submission pass.

Note: This restores rendering of all examples as the default material is used
if none is provided. On the other hand this should trigger warnings in the
console.

Change-Id: I20955bdc96b2844b2b33e46202c3e0e20634cc64
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>