Fix Deadlock issue
The Node has a QReadWriteLock and the QChangeArbiter a QMutex.

In the case where QChangeArbiter->registerObserver is called, the mutex is
locked and the observer Node's QReadWriteLock gets locked.
If at the same moment on the same Node within a different thread
notifyObservers is called, the QReadWriteLock gets locked and a call the
QChangeArbiter->sceneChangeEventWithLock is called which tries to lock the
mutex and eventually waits forever.

This scenario results in a DeadLock. It was fixed by unlocking the Node's QReadWriteLock
right after the QChangeArbiter is retrieved in notifyObserver so that sceneChangeEventWithLock
can be called with the Node's QReadWriteLock unlocked.

Change-Id: Ib3f6f5042239f3ad5a32daa0010aadbfe9fe3604
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>