Reordering order of creation of Jobs and QGraphicContext
QJobs for the current frame used to be created even if the QGraphicContext
hadn't been enabled. That meant that a few frames worth of RenderCommands were
created before having the QGraphicContext created. With the TechniqueFilter in
RenderView requiring a way to access OpenGL context properties to filter based
on the OpenGL version or extensions, this was reordered.

While the QGraphicContext hasn't been initialized, the Renderer returns no
Jobs to be computed. This ensures that RenderViewJobs will only be created
after the QGraphicContext so that they can access GL properties.

An interface exposing GL properties will be added in follow-up patches.

Change-Id: I3dea5f52fb97803f02c2f066037c999f85e02f74
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>