QGraphicContext queries shader uniforms and attributes
When a program shader is created, we retrieve the list of uniforms names and
locations and attributes names and locations.
This will allow us to efficiently see if Parameters provided by a Material are
needed/missing in a shader rather than relying on querying their location to
check their presence.

Change-Id: I611332ef2bd1d7c2bda50d30f0ec06cf059c3e3b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>