Move the bindinds of Uniform out of QUniformPack
QGraphicContext and QUniformValue now handle the bindings of uniforms with a
GL Shader.

A bit confused at the moment where StandardUniform should be defined. Should
we assume a documented name, should that be customizable through the shader.
Lots of pieces that could fit equally well in separated places. Trying to find
the best way to do so. Ideally we should assume that we could switch the
QGraphicContext to work with a different Graphics API. That forces to abstract
Uniforms, Attributes, Buffers, Shaders so that QGraphicContext is the only
class manipulating raw GL context.
Maybe rename QUniformPack to QParameterPack, StandardUniform to
StandardParameter ...

Change-Id: I84f175cf7c613183d70458da034847b9c7c6af83
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>