Improved Technique selection
RenderTechnique are now index by QAbstractEffect instead of
QAbstractTechnique.
That way we can detect with a RenderEffect and its peer if we have a
RenderTechnique associated or not. Also the technique filtering occurs only
when the RenderTechnique is created. This assumes that the TechniqueFilter
criteria have a constant value.

If we need to monitor TechniqueFilter for criteria updates, that requires
having a way to set the TechniqueFilter to dirty and implementing a locking
mechanism in RenderView to properly access that dirty flag.

Change-Id: Ia38060d6f8ba205a6df8083f24e6641ea35d0494
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>