Added ParameterPack
The goal is to have RenderMaterial, RenderEffect and RenderTechnique contain a
ParameterPack.
Each of the frontend Material, Effect and Technique should post a notification
when a Parameter is added or removed.
Upon reception of that notification in the backend classes should add the
Parameter to the ParameterPack. In turn, ParameterPack will register itself as
an observer of the Parameter and receive notification when the Parameter's
value is updated.

For that to work, Parameter should be converted to a Node so as to be able to
be registered as an observable.

Change-Id: I26373c8dc4006b4e41f2bcde8aec0d9ca751bc8a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>