Cleanup of Parameter
A Parameter is now used only for user specified uniforms. For StandardUniforms
and Attributes we could rely on the ParameterBinder. In case a mesh attribute
or standard uniform doesn't have the default name used by Qt3D shaders (that
will have to be specified somewhere), a ParameterBinder could be added to
solve that issue. Will see how that goes in a follow up patch.

Change-Id: I41bd0d4ef2ac731c109f87fca5b28571074bfba5
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>