QUniformValue templated subclass
QUniformValue is now a base class. A templated subclass contains the actual
uniform values. Having a template here allows to use the right  overloaded
setUniformValue method in QOpenGLShaderProgram.

Note: This commit is part of a series of modifications that need to be merged
together but are committed separately to ease the reviewing process.

Change-Id: I7c3e548ae33730a36a9a7ba83bbd281ba656e875
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>