Renderer refactored
Updated to not submit the modelMatrix and RenderCamera to the QGraphicContext
as not needed anymore. Also defined default handles to the default material
classes so that the defaultMaterial can be assigned directly when building
RenderCommands instead of checking when actually performing the rendering.

Note: This commit is part of a series of modifications that need to be merged
together but are committed separately to ease the reviewing process.

Change-Id: I5d7ade4b4e788e27a7add86221656de95c0418ba
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>