RenderShader updated
Doesn't contain standardUniforms any longer. StandardUniforms and user defined
Uniforms are now passed directly by a QUniformPack at rendering time.

A later improvement could be to check for a given shader, whether a uniform is
updated instead of resetting it every time.

Note: This commit is part of a series of modifications that need to be merged
together but are committed separately to ease the reviewing process.

Change-Id: Ib3028dfca42032e4bf37cd307bb4c87ba739845d
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>