simple-qml example updated
Updated to work with the latest versions.

Parameters are now used only for Uniforms. There is no longer a need to
specify the data type. That assumes on the otherhand that the shader uses the
correct type (QColor -> vec4 and not vec3).

StandardUniforms are specified by setting a StandardUniform enum as the
Parameter's value and setting a ParameterMapper with the StandardUniform binding type.

Attributes and now only defined by creating a ParameterMapper.

We could get rid of this level of configuration. However, I believe that we
can keep Material simples by using Parameter for used defined uniforms. Then
for standardUniforms we can define Parameters in Technique and/or Effect.
That way simple users will only touch Material and advanced users can have a
greater control with Effect, Technique, RenderPasses and ParameterBinder.

Note: This commit is part of a series of modifications that need to be merged
together but are committed separately to ease the reviewing process. This is
the last commit of the serie. Compiling and rendering should work from that
point on.

Change-Id: Ie2b3ae848e9c7e56fee685a0856bd58740bb0155
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>