Texture updated
Added Q_PROPERTY to be able to set all values from QML.
Couldn't register QOpenGLTexture enums directly as QOpenGLTexture is not a
QObject so enums values where copied to Texture. A static_cast allows to
convert directly a Texture enum to a QOpenGLTexture enum.
Corrected Quick3DTexture and registered it as an extended type.

Change-Id: I1a99a2dd549722cd647cc53f60e720e0c4039df9
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>