Renderer / RenderviewJobs : some ground work to use QFrameAllocator
Each RenderviewJobs is now assigned the index of the current frame to which it
is processing for. Is also now contains a QThreadStorage that contains a vector
of QFrameAllocator indexed by frame number. Later on, RenderView,
RenderCommand, QUniformValue ... will be allocated using the QFrameAllocator
specified for the current frame.
That way each Thread has its own set of allocators indexed by
frames and don't need locking to be accessed within the thread's context.
Allocators recycle their memory between frames which should help both
performances and reduce leaks.
The Renderer can now be constructed by specifying the number of
maximum cached frames to be processed.

Change-Id: I748bce37eb6dd9e60303bae8bed5db469d539139
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>