Rendering using QFrameAllocator
Renderer contains a QThreadStorage object containing a vector of
QFrameAllocator. Each Renderview Jobs calls the Renderer to retrieve the
QFrameAllocator for their thread and their frame index (between 0 and the maximum
number of cached frames). If the QThreadStorage hasn't been initialized, it is
initialized, otherwise the QFrameAllocator is returned directly. The
QFrameAllocator is automatically cleared if the previous call wasn't for the
same frame index.
QUniformValues and RenderViews have been updated to used QFrameAllocator to
allocate objects instead of using new.
RenderQueue has been slightly modified. It is now possible to retrieve the
RenderQueue to render without popping it out of the QCircularBuffer of
RenderQueues. A call to popFrameQueue does that. This is needed to be sure
that we won't be rendering a RenderQueue that could have been cleared because
the backend is rebuilding a newer frame. This forces the Renderer to render
the whole frame before allowing the Jobs to build other frames if there are
more than the maximum amount of cached frames.

Change-Id: Ie7718e201e2f6b420903092d1706277e9786916e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>