lightProperties method in lights
QAbstract implements a pure virtual lightProperties methods.
This forces subclasses to implement a method returning a hash of variant
properties indexed by their name. This has to be done in order to have a
single RenderLight element monitor all types of lights and to allow future
addition of new light types.

Change-Id: I5124ee34f75abe80992ee910cb0da592561fde5c
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>