RenderCommand / Renderer prepartion for automatic uniform bindings
RenderShader now has a hash of uniform and a hash of attributes both
containing  names and glsl binding positions.
Those are initialized when the ShaderProgram is created by the
QGraphicContext.
In order to initialize uniforms and attributes only once the shader has been
created, RenderCommand and Renderer have been slightly modified.

Change-Id: I85b8769195a81834f3919f18a9f53cbf5e032ddb
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>