QAspectManager slightly reworked
The QAspectManager has a getter to the window.
registerAspect now initializes directly the aspect prior to registration.
The aspect is at that point able to retrieve the window through the
QAspectManager if needed.
The setRoot method now only calls registerAspect on the aspects.
The setWindow method was removed from all aspects

Change-Id: I730861a089be15220f7fd36201d15d59ed3ede70
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>