Renderer QGraphicsContext initialized at startup
RendererAspect, RenderThread and Renderer have been modified to create the
QGraphicsContext prior to setting the SceneGraph root node. This will allow to
query the QGraphicsContext when building the FrameGraph/SceneGraph trees.
Then RendererAspect launches and waits for the RenderThread to be running, all
threads are locked at that point (but the QML thread on which we have no control).
The RenderThread creates the Renderer and puts it in a waiting for
initialization state prior to allowing any other thread to continue their
execution. When the RendererAspect assigns a Window to the Renderer, this
wakes up the Renderer which was waiting for initialization. Initialization is
therefore performed and the QGraphicContext is created. Note that a call to
QWindow::create is performed prior to that so that the window has its storage
allocated eventhough it's not yet being shown. The issue we had previously with
QGraphicContext creation failure was also caused by that, in most cases
QWindow::show (which calls create) was called before we tried to create the
QGraphicContext but eventually this would not always be the case, hence some
failures. Once the window and the QGraphicsContext have been initialized, the
setSceneGraphRoot of the Renderer is called and frame processing is launched.

Change-Id: Ib18faef5b38d88fdf09fe6b27940af9ad0a1cd6e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>