QAbstractMeshFunctor
Each QAbstractMesh class now has to implement a QAbstractMeshFunctor class.
When a mesh is set to dirty, the functor is sent through a QChangeArbiter
notification to the backend. The backend, using the functor can then rebuild
the mesh. There should be no issue if the frontend mesh is deleted while the
backend is creating the mesh as the functor has no direct reference to the
frontend element and contains all the data needed for a complete creation of
the mesh.

Change-Id: I4984f03a612e74c688bfb6cc2f19d9241b517457
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>