Ground work to have a proper shutdown
QAspectEngine shutdown invokes QAspectManager quit which will stop the
QAspectThread. Prior to exiting, the QAspectThread calls QAspectManager
shutdown which cleans up all aspects and deletes them.
Renderer stops the RenderThread when cleanup is called and properly deletes
the QGraphicContext and the RenderThread.
Note: Everything is called as it should. However it always ends up in a
segfault caused by the QFont::cleanup when the QThreadStorage of the main
thread GuiApplication is destroyed. I've not been able to find the cause of
that.

Change-Id: I3b7baa843af4d0418b5d04e6d1cda88b962ded7f
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>