RenderEntity refactoring
Components are indexed by their own uuid instead of by the Entity::uuid.
The backend tree is therefore smaller as we only have one RenderMesh per
Mesh instead of having a RenderMesh per RenderEntity.

RenderEntity also contains a set of templated helper method that can be used
by the RenderSceneBuilder or other classes that need to add a component to a
backend RenderEntity. Applying the same thing to RenderMaterial / RenderEffect
would provide a cleaner code.

Change-Id: I36cf76ed717fa58275c66cc20d1e7732579c35fa
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>