simple-qml update to use RenderPassFiltering
The top right viewport now has a RenderPass filtering with criterion search
for RenderPasses named Texture. This has for result of showing only the
textured ball and use the default renderpass for other meshes.
In the bottom viewport, no filtering is used and balls that are rendered
without effect in the top right viewport, should appear with a changing color.

Change-Id: Id7c6c497dae93433381bb116608239539ffac78a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>