RenderPassFilterNode
Uses uuid to register peer.
Stores uuid of criteria instead of handles as the handle for a criterion may
not have yet been created when a RenderPassFilterNode is created. As we
traverse the tree from root to leaf, children contained by a
RenderPassFilterNode are only created after the RenderPassFilterNode itself.
A the criterion quuid allows to reference criterion that will be created later.

Change-Id: I1f79d11f2998dc7b2c7ec3743629c7e402d98137
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>