QResourcesManager bug fixes
Found out that not naming the ReadLocker / WriteLocker caused them to be
directly destroyed hence not locking anything.
If some resource is acquired with acquire (not using an indexing id) it must
explicitly be released using release and the handle. This allows to make
release significantly faster by not having to lookup if there is an id
associated with a handle.
ListAllocatingPolicity slighty improved. However there seem to be an issue
with ListAllocationgPolicy where using a QList might be, by its nature a bad
idea. If the QList is reallocated when it needs to be grown, all pointer values
returned by allocate may be invalid and then cause a crash. We'll have to see
what we can do for that.

Change-Id: I049d322c02e686ee2e97928a6fd8b648be5d2e62
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>