Deallocate RenderView and RenderCommands when rendered
The operator delete was overridden in RenderView in order to deallocate the
RenderView, RenderCommands and other resources from the QFrameAllocator when
the renderer deletes the list of RenderViews.

Without this, having a QFrameAllocator is useless as the QFrameAllocator
would keep on allocating new memory chunks as no resource is ever released.

Change-Id: Iffbe546f23532ddcf424dce6164c93020dc2e617
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>