Add QOpenGLDebugLogger integration to Renderer
When initializing the renderer, check whether the environment variable
QT3D_DEBUG_LOGGING is set. If so, ask for a debug OpenGL context and if
the GL_KHR_debug extension is present use QOpenGLDebugLogger to start
logging OpenGL debug info.

If QT3D_DEBUG_LOGGING is set to a string starting with "sync" the logging
will be synchronous. The string comparison is done in a case independent
manner.

Change-Id: Ia29e12abfd593336b0304ea37aa79285efca06c6
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>