QSceneObserverInterface sceneNodeUpdated method for global notifications
This method will be called with a corresponding QSceneChangePtr when the
backend or any other aspect wants to update a particular frontend Node.

A frontend scene observer will be created, living in the man thread, it
will invoke a method that will send the update to the node.

This will allow to either have all aspects update themselves directly or
update only the frontend node and make it notify all backend aspects.

Change-Id: I492e707a6901c204f0c4aefa88b4d581e47cf496
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>