QAspectEngine QNode lookup table
QNode now contains an instance of the QAspectEngine. When the scene is first
created, all nodes are traversed and inserted in a table referenced by their
uuid.
When new nodes are added at runtine, QNode::addChild takes care of inserting
the new node into the QAspectEngine's table. The same applies for
QNode::removeChild.
QNode now also has a isCloned() method.
QFrontendSceneObserver has a pointer to the QAspectEngine.
QFrontendSceneObserver will need to check in the QAspectEngine's table if a
node is available to receive the notification when sceneUpdated is called.

Change-Id: Ic243261ecff5e15f841ca180cb97ba65df2b3bfb
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>