Improved FBO initialization
Initializes FBO and setAttachments and drawBuffers once.

QTexture/RenderTexture need to be reworked to handle texture size changes
which could be useful when resizing the window in a deferred rendering
context.

Also allow to render meshes even if there is no camera, this is the case for a
full screen quad in a deferred renderer for instance. Corrected direct camera
access accordingly.

What's really missing now is a glBindFragDataLocation.

Change-Id: I2e1b1141a6367a5a63b9bd01da44e3b938b7758b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>