Add a wireframe example to exercise geometry shader support
Needed to extend the set of standard uniforms to include the viewport
matrix.

We also don't seem to get the type information for uniforms from glsl
but rather infer it from QML. In this example it resulted in a bug
where the line.width was being set with glUniformi() instead of
glUniformf() when we used whole values such as 1.0 in the Parameter of
WireframeEffect.qml. Worked aroudn by using a typed property in
WireframeMaterial.qml for now but this needs fixing.

Change-Id: Ie07a81b1b0b6c176e188d9e489643b22e339192a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>