AssimParser slightly refactored
It turns out that Assimp doesn't export the dtor of aiScene.
We used Importer::GetOrphanedScene to manage ourselved the
lifetime of aiScene. However as the dtor is not exported,
changes had to be made to let Assimp manage the lifetime
of aiScene.

It was probably working before because I had a local build of
Assimp but, the issue appears with distro packages.

Change-Id: I14f2b5c7bae1555961bc47fa335d54c2f0bc969a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>