Always break out of the renderer's main loop when renderer is destroyed
It was possible that the renderthread could be waiting on the RenderView
population when we were asked to terminate resulting in a deadlock. We
now explicitly wake up the wait condition and then check for termination
criteria after waking up.

This is the first step towards having a clean application shutdown.

Change-Id: I4147843f658ae563798bcdad4d308cd5c641c9e9
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>