Prevent deadlock during application shutdown
There is a potential deadlock when shutting down. The call to
QAspectManager::quit() sets the atomic ints such that they break out of
the loops in void QAspectManager::exec(). However, without this change
the inner loop gets one last iteration where it asks the aspects for
jobs to execute. In the case where the frame queue of the renderer is
already full, this blocks and if the renderer has already stopped
processing frames from the frame queue this will block forever.

Moving the processEvents() call after the job submission prevents this
by allowing the loop to exit immediately and avoiding the deadlock.

There is still more to be done as there is some crash around a mutex
still whilst destroying some other objects.

Change-Id: Iac5159095e5714dc1ba8c9ce128cb3a5aaaec90b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>