RendererAspect: register backend classes with frontend types
RenderSceneBuilder was removed, this is node replaced by NodeVisitor
traversing the scene add using the backend functors to create elements.

Added a QFrameGraph functor so that we can easily set the active framegraph in
the Renderer. Ideally, QFrameGraph should have a proper backend element and we
would need a way to find all the components attached to a RenderEntity.

Change-Id: I45d51fee63c4f07aaa6c7af297a7fbe9c71254a0
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>