Optimize ArrayPreallocationPolicy.
The tricks applied are identical to those done previously for
ArrayAllocationPolicy:

We remove the obsolete hash for the index mapping and instead use
pointer differences directly, which is much faster and reduces
the memory consumption.

Some variables are also renamed to better reflect common allocator
nomenclature.

The added benchmark shows a reduction of the instruction count
by a factor of about 10.

Change-Id: Icabc354ab9e1e8ea0b1dc83c71a2f2cfe49d1da0
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>