Optimize QHash usage in RenderShader::updateUniforms.
It allocated a temporary QList of keys and repeatedly queried the
items in the hash, incurring relatively slow QString hashing.

Using iterators instead we can get rid of the temporary QList and
safe the hash computations on the outer list.

Profiling with Linux perf shows that the relative cost of this
function drops from around 2.0% to less than 0.5%.

Change-Id: Ic6a3c3ed7b6b4242e763a0b21987d543021de332
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>