Fix RenderTransform::sceneChangeEvent
Patch 54d0ef3ba (Micro-optimize RenderTransform::sceneChangeEvent) removed the
check against the property name in RenderTransform and considered that the
NodeUpdate notification would always be for the"matrix" property. This is
unfortunately not the case and breaks the shadow-map-example. We receive
notifications for all user defined properties that are defined within a QML
Transform element.

This patch restores the property name check.

Change-Id: I50ec3aab2e982d5a6055c73c3153211f4042669f
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>