QKeyboardInput: add focus property
Like Item focus in QtQuick the goal is to change the current QKeyboardInput
focus for a given QKeyboardController when focus is set to true. Only a single
QKeyboardInput for a given QKeyboardController can have its focus set to true
at a given time. Setting focus to true should requests a focus change to the
backend.

Change-Id: Idaa42192076a6b67221fb112deb53642bfc2dd93
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>