Aspects: onInitialize now takes a QVariantMap
The QAspectEngine setSurface method was replaced by a setData method which
takes a QVariantMap. This offers a way to specify different values that
different aspects may be interested in. It is the responsibility of the Aspect
subclass to check that the QVariantMap contains the necessary element when
onInitialize is called.

Change-Id: Ifcf5d2ee9e5fac7b0fe3c320c1415bcf55f1cf55
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>