Guard against unsafe usage of QFrameAllocator::(de)allocateRawMemory.
Before one could potentially fall into a trap when one would call

allocateRawMemory<BaseClass>(size);

When size belongs to a child class inheriting from BaseClass and
falls into a different (larger) bucket, the following deallocation
via

deallocateRawMemory<BaseClass>()

would fail. Because there we just picked the size of the BaseClass
to find the bucket.

Now the whole API is restructured to guard against this wrong usage:

- we only ever operator on void* in the *Raw* methods
- we pass on the size explictily

This also means we must use the safe operator deallocate overloads
that also take the size as a parameter.

The unit test was run first to check that this was indeed an issue.
Then it is ported to the new API and cleanup up a bit as well to
reduce duplicated code.

Change-Id: I559af1c0ffc633962fac12d6e3cadf87c09ef92d
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>