Cleanly exit the render thread
We have to mutex protect accesses to the m_surface variable in the
Renderer as the main thread may call destroy() on the corresponding
QWindow. This can happen in response to the window being closed or
changing QScreens.

Calling swapBuffers() on a QSurface that has been destroyed does not
fail gracefully but crashes. This change works around this by using
the new QPlatformSurfaceEvent in QtGui which is sent to the window
or offsceen surface when the platform surface has been created or
is about to be destroyed.

This event is used to inform the renderer to stop rendering to the
window surface in a safe manner, before the native surface gets
destroyed.

The Renderer::submitRenderViews() now checks at the start of each
frame that the m_surface is still valid.

There is still another crash at shutdown if applications are
sending a lot of updates to the backend.

Change-Id: I742226c7ae8f657e218cbcd45f0dff3d2b7bcc18
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>