Rename AxisAlignedBoundingBox->QAxisAlignedBoundingBox
We may be able to make this a private class so as to not use up the
name for a type we may need to export to QML in order to allow the user
to specify a custom bounding box for use with culling if the shader they
are using does some non-standard scaling (e.g. instancing).

Change-Id: Id7120eba6407060507ac3d4c45061ce9670557c7
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>