Change a warning into debug output
It doesn't really matter if a mesh has an attribute the shader doesn't
use. It's just a bit of wasted bandwidth but not really worthy of a
warning and it almost always happens with the shipped examples.

Perhaps we need a way of saying which attributes get loaded/generated
from obj files and similar.

Change-Id: I616acd9f8cbc3e403359609f7fe9aa5987612d59
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>