Optimize: Prefer QVector over QList for QObserverList.
This list stores QObserverPair, i.e. a pair of a QFlags and pointer,
which by definition is larger than a pointer. Thus QList allocates
each node on the heap, whereas a QVector can use a single block
of memory.

This reduces allocations and improves iteration speed, as this reduces
the cache misses incurred with QList and pointer chasing.

Change-Id: Ib3a9fbaa099cb41785086857e856072b74211605
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>