Quick3DShaderDataArray
QML allows property var array: [] of simple types only.
We can't use a var property to store an array of ShaderDataObject so
Quick3DShaderDataArray was introduce to overcome that issue.
It contains a QQmlListProperty<Quick3DShaderData>.

This is only needed for QML, in C++ providing a QVariantList of QShaderData is
enough.

Change-Id: I768f5189252c219c5a1718b4c4da128bbaff008e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>