Fix parameter priority regression
Commit e7443372bc036839c020d372265d4708d2eca4d7 accidentally introduced
a regression in the construction of the list of parameters for a shader.

In that commit addParametersForIds() was changed to not insert a
parameter if a parameter with the same name was already in the list.
This changed the semantics of the function but the calling order wasn't
updated accordingly.

This fixes the regression by filling the list of parameters in the
correct priority order.

Change-Id: I5a33a137c40fe463d42e8459dc197b295f248e63
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>