Fix dynamic reparenting issues
Changing the parent of an Entity at runtime caused lots of crashes.
These are a few notes / changes that allow the program not to crash during
such changes.

* QEntity dtor calls removeAllComponents only on non clone entities
* We do not clone the parent on child but the child directly when sending a
  notification about a node removed.

I suspect the EntityLoader is not deleting the previously loaded Entity in the
loader-qml example and therefore didn't produce a crash.
This might have also impacted the shutdown of any Qt3D application.

dynamicscene example to test that in a follow up patch.

Change-Id: I4ff62ae670e35ef6b783b4fc5babec942c082c9a
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>