Move CameraController in the input aspect
With this move it also becomes a private class. As a result I had to
touch all the examples as you won't get camera navigation anymore if the
input aspect is not registered.

Quick3DConfiguration doesn't depend on the QuickWindow singleton hack
anymore, it tries to get back to the aspects and set the camera on the
ones having a corresponding property.

It also mostly empties Window and QuickWindow.

Change-Id: Ife1644bd338e51929bad1e99089bd5e4c68ebc68
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>